refactor(reward-partner): update DTOs
