+++ src/reward-partner/infrastructure/dtos/login-reward-partner.dto.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { emailDTO, passwordDTO, usernameDTO } from './dto-types/dto-types'
+import { usernameDTO, emailDTO, passwordDTO } from '../../../shared/infrastructure/dto-types/dto-types'
 
 const LoginRewardPartnerDTO = z.object({
   username: usernameDTO.optional(),
