+++ src/reward-partner/infrastructure/dtos/register-reward-partner.dto.ts
@@ -1,5 +1,14 @@
 import { z } from 'zod'
-import { addressDTO, cuitDTO, emailDTO, nameDTO, passwordDTO, phoneNumberDTO, usernameDTO } from './dto-types/dto-types'
+import {
+  addressDTO,
+  coordinatesDTO,
+  cuitDTO,
+  emailDTO,
+  nameDTO,
+  passwordDTO,
+  phoneNumberDTO,
+  usernameDTO
+} from '../../../shared/infrastructure/dto-types/dto-types'
 
 const RegisterRewardPartnerDTO = z.object({
   name: nameDTO,
@@ -8,7 +17,8 @@ const RegisterRewardPartnerDTO = z.object({
   cuit: cuitDTO,
   email: emailDTO,
   password: passwordDTO,
-  phoneNumber: phoneNumberDTO
+  phoneNumber: phoneNumberDTO,
+  coordinates: coordinatesDTO
 })
 
 export default RegisterRewardPartnerDTO
