+++ src/reward-partner/infrastructure/dtos/update-reward-partner.dto.ts
@@ -1,11 +1,11 @@
 import { z } from 'zod'
-import { addressDTO, emailDTO, nameDTO } from './dto-types/dto-types'
+import { nameDTO, addressDTO, emailDTO, phoneNumberDTO } from '../../../shared/infrastructure/dto-types/dto-types'
 
 const UpdateRewardPartnerDTO = z.object({
   name: nameDTO.optional(),
   address: addressDTO.optional(),
   email: emailDTO.optional(),
-  phoneNumber: z.string().optional()
+  phoneNumber: phoneNumberDTO.optional()
 })
 
 export default UpdateRewardPartnerDTO
